requirements.txt
opencv-python>=4.8
mediapipe>=0.10
numpy>=1.24
pydantic>=2.0
pyyaml>=6.0
simpleaudio>=1.0
